Break down app file into function module and router file.
